Time: 5 ms (80.48%), Space: 22.7 MB (97.1%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -19,7 +19,7 @@ public:
         while (left <= right)
         {
             int mid = left + (right - left) / 2;
-            int time = 0;
+            long long time = 0;
             for (int x : piles)
             {
                 time += (x + mid - 1) / mid;
